Current design doesn't support embedding
Major refactor to encapsulate most of the core PaySim logic into
an abstract class. Plan is to support easier embedding in 3rd
party code, specifically something that can feed transactions into
a database like Neo4j as the simulation runs and not have to worry
about accumulating them and writing them out to disk.

Also tried to do some minor cleanup along the way...mainly types.

New abstract interface revolves around two abstract methods that
work basically as even handlers:

- boolean onStep(long stepNum)
- void onTransactions(List<Transaction> transactions)

The onStep is called after each step iteration with the current
step number. Its return boolean return value is used to determine
if the simulation should abort.

The onTransactions method is called whenever transactions are
recorded in the simulation. It's up to the implementing class to
decide to hold onto this state or do something immediately with it.
Since I believe this is all single threaded in how the transaction
is ultimately orchestrated, it's a blocking process at the moment.