Transactions only imply actor types
This refactor establishes a formal SuperActor "Type" (as an enum). It
exposes SuperActor as an abstract base class now forcing all
implementors to provide a getType() method.

In the case of a Mule, it's a special instance of a Client. This allows
the Mule to override the getType() to identify itself as such (without
dealing with reflection or class inspection) yet keep its "cover" of
using the "C" prefixed name in the original logs.

Transaction was updated to have a constructor based on SuperActors
instead of just names and will now also grab the Type. It's a candidate
for a further refactor to reduce code by auto calculating balance
changes which are currently scattered throughout Client methods it
seems.