Need a way to test core PaySim logic
This commit addresses a few design things, but specifically refactors to
make it easier to test a "gold standard" data set for some known inputs.

Since PaySim is deterministic, and we assume the original PaySim
implementation is "correct," memorialize the output of that known-good
PaySim with some settings that make for a small-enough output to safely
stick in git.

The refactor adds a worker thread to the IteratingPaySim to offload
running the simulation in a separate thread. Data is communicated via a
BlockingQueue, exposed to the user as an implementation of Iterable.

Some design issues exist since implementing something like hasNext() is
challenging due to not being able to just look for an empty queue since
the worker generates data async. Currently, calling next() will block
the calling thread until there's data in the queue...which effectively
waits for the sim to produce data.