Parameters being global prevents running multiple sims
Refactor to make the Parameters class not have a bunch of 'public
static' fields. Makes the API a little clunkier for now (in the actors),
but not too bad. Parameters should be fixed for the duration of a
simulation, so having a sim class keep a reference to a copy is fine.

Parameter fields are directly exposed, but now set to final so we don't
have to add a bunch of getters. (yuck)

The properties file still depends on the goofy hiearchy of files though.