Can't abort a simulation, hangs in embedded scenarios
Wire in a hacky abort() method for the IteratingPaySim implementation.
Not completely sure it's 100% fail-safe, but should help cleanly abort
and end the worker thread.

Add a simple test to the abort routine.

To help testing, this commit also adds the ability to configure the
queue depth used as the transaction buffer.