Can't tell order of transactions per Client
This commit introduces a global step counter under the following
assumptions:

1. The simulation is deterministic in sequence and will always produce
the same transactions in the same order for a given seed value.
(Effectively it seems we can consider the simulation having a single
producer.)

2. While knowing global transaction order is of questionable value at
the moment, knowing transaction order at the Client level where a Client
may perform multiple transactions per step (equivalent currently to an
hour of real time) is valuable for reproducing event chains.

For safety, the global counter is implemented as an AtomicInteger (from
the java.util.concurrent.atomic package). For now it's only implemented
in the IteratingPaySim implementation. The global step value is also
added as an attribute to the Transaction class with getter/setter.

Future commits may push the global step counter down into the base
simulation class.

Lastly, crank the version as well.