Writing to stdout directly is annoying for libraries
The project already uses SL4j, so this commit migrates to using the
Logger interface provided by the library.

For now, it also bundles the sl4j-simple implementation, so a future
commit may properly refactor this project into a library version (w/o
the implementation) vs a standalone "app" version.