Need a way to start modeling 1st/3rd party fraud
PaySim originally lacks an idea of shared identifiers. If it happens,
it's accidental. This commit starts to incorporate means of generating
both identifiers (formerly the name attributes of the actors) and some
"recognizable" name (like "Bank of XYZ" or "John Doe").

Next step will be to wire in some logic to make for some shared
identifiers as this commit only starts to ensure they're globally unique
for a simulation.