Two big changes: uniqueness & history
Problem 1: ids weren't being generated to be unique.
This commit makes sure that by default all id's are globally unique for
Clients, Merchants, and Banks.

Problem 2: Fraudsters were victimizing too randomly
Adds a BoundedArrayQueue, which acts as basically a sliding buffer, so
actors can remember their last N transaction partners. Primary usage is
to peek into a Merchant's recent history to find a Client to victimize.

On that note, Fraudsters now initialize and pick potential Merchants
from the universe to use when seeking victims. For now, this implements
a naive algorithm:

- At initialization, pick 3 Merchants at random
- When it's the Fraudster's turn to act, perform the fraud probability
  check per normal. If it fails, abort.
- 80% of the time the Fraudster commits fraud, pick a random Merchant
  and then pick a random Client from their recent client history. They
  become the lucky victim.
- The other 20% of the time, pick a Client from the universe at random.

In addition, victims are now tracked as well, and the fraud reporting
now includes the client id's of each victim.