More identifiers, more focused fraud
Add in some initial, crude support for adding arbitrary properties to
actors. In the process, add emails and phone numbers to Clients.

Make a "high risk" pool of merchants to base fraudster faves off
of...cleanup some of the code nearby to prevent recalculation of
population sizes to minimize chance of logic errors in future.