jfairy too tightly coupled, identities messy
This pulls the jFairy stuff out into an identity package to help
centralize its usage and buffer all the actor code from knowing about
jfairy details. Plus, this moves some of the logic out of the
PaySimState abstract class, too.

Identities are still a mess among the actors as they primarily only
pertain to Clients...and the Client class's contstructor is ripe for
some cleaning.

Next steps are to:
- expose identities for all Client implementations
- add identity spoofing to 1stPartyFraudster
- clean up the constructor mess