Constructors are a mess using diff'nt conventions
First step at simplyfing the constructors to the actor types...not going
to go much further in OOP design here, but now that "name" is moving
into an "Identity" we can simplify things. In surveying the code, id
generation was always going through the PaySimState instance anyways, so
making it so we just pass a reference to the simulation state cleans up
some of the code.

There is known copy-pasta in some contructors, mainly the Client and
Fraudster types, dealing with setting up base properties from the given
or generated Identity. Whatever...this is fine for now.

The design goal here is to keep the general PaySim structure of focusing
on Actors and Transactions, but carry through enough metadata to create
things like extra Identities and relationships in the demo code.