Fix #1 - Clients ended up with multiple SSN assigned
Bit kludgey...changes logic to generate a brand new "unique" identity
that it then mutates to generate the synthetic one using existing
identifiers/identities. This removes the re-use of the same "base"
ClientIdentity that was resulting in Clients with multiple :HAS_SSN
relationships in the downline paysim-demo.

The added test is pretty ignorant and low value...just checks for
the faux identities have colliding Client ids.