v2.1.3 fix 3rd party fraudster logic not triggering
Short story: using the transfer limit as a way to design a "test"
charge was flawed from the get-go. These changes make the test
charge based on the Client's assigned average value for Payment
transactions, which resolves the issue where every test charge was
basically "failing" because it was quite often well over the
balance of the victim's account.

Does not update Windows test set, that will be a separate commit.